XWIKI-12922: The flavor picker is broken.
The first line of the resource was a velocity comment. Once reduced by the CSS minifier, the whole file was commented!
This fix updates the comment by using a CSS comment with a start and an end. We could have done the same with a velocity comment,
but it seems more logical to use a CSS comment in a CSS file!

(cherry picked from commit 6df9d57d9ae29c18b6f66629c1a77bbfaca75f9f)